+++ ChangeLog.md
@@ -2,6 +2,7 @@
 ## 4.4 - 2022-07-
 
 - rename of clusters in the cluster editor will be immediately reflected in rendered graph
+- Package Analysis: reporting skipped assemblies fixed
 
 ## 4.3 - 2022-06-27